% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\docType{data}
\name{cj_map_theme}
\alias{cj_map_theme}
\title{Default ggplot themes}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 93.
}
\usage{
cj_map_theme
}
\arguments{
\item{No}{params}
}
\value{
ggplot2 themes
}
\description{
Clean map theme and plot theme using serif font.
}
\keyword{datasets}
