% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\docType{data}
\name{cj_plot_theme}
\alias{cj_plot_theme}
\title{Default ggplot plot theme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 93.
}
\usage{
cj_plot_theme
}
\arguments{
\item{No}{params}
}
\value{
ggplot2 theme
}
\description{
Clean map theme.
}
\keyword{datasets}
