% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_chunk_opts.R
\name{default_chunk_opts}
\alias{default_chunk_opts}
\title{Set default knitr chunk options}
\usage{
default_chunk_opts(cache = FALSE)
}
\arguments{
\item{cache}{option}
}
\value{
nothing
}
\description{
Center figures with default latex positioning, allow custom text
Allow choosing whether to use cache, set custom code block size
and colour. Dont show warnings, code chunk size width set.
}
