% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_opts.R
\name{default_latex_chunk_opts}
\alias{default_latex_chunk_opts}
\title{Set default knitr chunk options for latex}
\usage{
default_latex_chunk_opts(cache = FALSE, ...)
}
\arguments{
\item{cache}{option}
}
\value{
nothing
}
\description{
Set more sensible chunk options for LaTex.
}
