% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_html_table}
\alias{make_html_table}
\title{Default html table options}
\usage{
make_html_table(
  df,
  caption = "",
  digits = 2,
  col_names = NA,
  row_names = NA,
  align = c("l", rep("c", ncol(df) - 1)),
  ...
)
}
\arguments{
\item{data.frame}{and various kable options}
}
\value{
kable table
}
\description{
HTML version of latex table
}
