% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_latex_table.R
\name{make_latex_table}
\alias{make_latex_table}
\title{Default latex table options}
\usage{
make_latex_table(
  df,
  cap = "",
  dig = 2,
  col_names = NA,
  table_env = "table",
  ...
)
}
\arguments{
\item{data.frame}{and various kable options}
}
\value{
kable table
}
\description{
Set nicer kable defaults, number of digits, customise table environment,
e.g. use *figure to make table go across two columns. Add custom styling
for bold headings.
}
