% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_opts.R
\name{default_latex_chunk_opts}
\alias{default_latex_chunk_opts}
\title{Set default knitr chunk options}
\usage{
default_latex_chunk_opts(cache = FALSE, ...)
}
\arguments{
\item{cache}{option}
}
\value{
nothing
}
\description{
Set more sensible chunk options for LaTex:
fig.align = "center", fig.pos = "tb",
fig.showtext = TRUE,
cache = cache, autodep = FALSE, echo = FALSE,
cache.path = "data/cache/",
message = FALSE, size = "footnotesize",
background = rgb(0.97, 0.97, 0.97),
warning = FALSE,
comment = NA, out.width = ".75\\linewidth"
}
